+++ Entrevistas.Domain/Entities/FatoresConversao.cs
@@ -5,8 +5,8 @@ public class FatoresConversao
     public string TipoEntrada { get; set; } = null!; // ScaleFactor ou EffortMultiplier
     public string Contexto { get; set; } = null!;    // Ex: PREC, RUSE, etc.
     public string Nivel { get; set; } = null!;
-    public double FatorConversao { get; set; } 
+    public decimal FatorConversao { get; set; } 
 
-    public string? NomeCompleto { get; set; }        // Novo campo
-    public string? Descricao { get; set; }           // Novo campo
+    public string? NomeCompleto { get; set; }
+    public string? Descricao { get; set; }
 }
